fixes minor build problems

+++ src/tamm/types.hpp
@@ -26,7 +26,7 @@ struct SpinSpace;
 using Spin = StrongNum<SpinSpace, uint32_t>;
 struct SpatialSpace;
 using Spatial    = StrongNum<SpatialSpace, uint32_t>;
-using TensorRank = uint32_t;
+using TensorRank = size_t;
 struct OffsetSpace;
 using Offset = StrongNum<OffsetSpace, uint64_t>;
 struct BlockIndexSpace;
